整合 GetAccessToken 和 Login 接口 到 auth.go
